+++ tests/activation_test.py
@@ -1,16 +1,17 @@
 import jax.numpy as jnp
 import pytest
 from e3nn_jax import Irreps, IrrepsArray, scalar_activation
+import e3nn_jax as e3nn
 
 
 def test_errors(keys):
-    x = IrrepsArray.randn("0e + 1e", keys[0])
+    x = e3nn.normal("0e + 1e", keys[0], ())
     with pytest.raises(ValueError):
         scalar_activation(x, [None, jnp.tanh])
 
 
 def test_zero_in_zero():
-    x = IrrepsArray.from_any("0e + 0o + 0o + 0e", [jnp.ones((1, 1)), None, None, None])
+    x = IrrepsArray.from_list("0e + 0o + 0o + 0e", [jnp.ones((1, 1)), None, None, None], ())
     y = scalar_activation(x, [jnp.tanh, jnp.tanh, lambda x: x**2, jnp.cos])
 
     assert y.irreps == Irreps("0e + 0o + 0e + 0e")