+++ tests/batchnorm_test.py
@@ -1,18 +1,16 @@
-import jax
 import jax.numpy as jnp
 import haiku as hk
-from e3nn_jax import BatchNorm, Irreps
 import e3nn_jax as e3nn
 from e3nn_jax.util.test import assert_equivariant
 import pytest
 
 
-@pytest.mark.parametrize("irreps", [Irreps("3x0e + 3x0o + 4x1e"), Irreps("3x0o + 3x0e + 4x1e")])
+@pytest.mark.parametrize("irreps", [e3nn.Irreps("3x0e + 3x0o + 4x1e"), e3nn.Irreps("3x0o + 3x0e + 4x1e")])
 def test_equivariant(keys, irreps):
     @hk.without_apply_rng
     @hk.transform_with_state
     def b(x, is_training=True):
-        m = BatchNorm(irreps=irreps)
+        m = e3nn.BatchNorm(irreps=irreps)
         return m(x, is_training)
 
     params, state = b.init(next(keys), e3nn.normal(irreps, next(keys), (16,)))
@@ -30,22 +28,22 @@ def test_equivariant(keys, irreps):
 @pytest.mark.parametrize("normalization", ["norm", "component"])
 @pytest.mark.parametrize("instance", [True, False])
 def test_modes(keys, affine, reduce, normalization, instance):
-    irreps = Irreps("10x0e + 5x1e")
+    irreps = e3nn.Irreps("10x0e + 5x1e")
 
     @hk.without_apply_rng
     @hk.transform_with_state
     def b(x, is_training=True):
-        m = BatchNorm(irreps=irreps, affine=affine, reduce=reduce, normalization=normalization, instance=instance)
+        m = e3nn.BatchNorm(irreps=irreps, affine=affine, reduce=reduce, normalization=normalization, instance=instance)
         return m(x, is_training)
 
-    params, state = b.init(next(keys), irreps.randn(next(keys), (16, -1)))
+    params, state = b.init(next(keys), e3nn.normal(irreps, next(keys), (20, 20)))
 
     m_train = lambda x: b.apply(params, state, x)[0]
     m_eval = lambda x: b.apply(params, state, x, is_training=False)[0]
 
-    m_train(irreps.randn(next(keys), (20, 20, -1)))
+    m_train(e3nn.normal(irreps, next(keys), (20, 20)))
 
-    m_eval(irreps.randn(next(keys), (20, 20, -1)))
+    m_eval(e3nn.normal(irreps, next(keys), (20, 20)))
 
 
 @pytest.mark.parametrize("instance", [True, False])
@@ -54,17 +52,17 @@ def test_normalization(keys, instance):
     sqrt_float_tolerance = jnp.sqrt(float_tolerance)
 
     batch, n = 20, 20
-    irreps = Irreps("3x0e + 4x1e")
+    irreps = e3nn.Irreps("3x0e + 4x1e")
 
     @hk.without_apply_rng
     @hk.transform_with_state
     def b(x, is_training=True):
-        m = BatchNorm(irreps=irreps, normalization="norm", instance=instance)
+        m = e3nn.BatchNorm(irreps=irreps, normalization="norm", instance=instance)
         return m(x, is_training)
 
-    params, state = b.init(next(keys), irreps.randn(next(keys), (16, -1)))
+    params, state = b.init(next(keys), e3nn.normal(irreps, next(keys), (16,)))
 
-    x = jax.random.normal(next(keys), (batch, n, irreps.dim)) * 5 + 10
+    x = e3nn.normal(irreps, next(keys), (batch, n)) * 5
     x, state = b.apply(params, state, x)
 
     a = x.list[0]  # [batch, space, mul, 1]
@@ -77,12 +75,12 @@ def test_normalization(keys, instance):
     @hk.without_apply_rng
     @hk.transform_with_state
     def b(x, is_training=True):
-        m = BatchNorm(irreps=irreps, normalization="component", instance=instance)
+        m = e3nn.BatchNorm(irreps=irreps, normalization="component", instance=instance)
         return m(x, is_training)
 
-    params, state = b.init(next(keys), irreps.randn(next(keys), (16, -1)))
+    params, state = b.init(next(keys), e3nn.normal(irreps, next(keys), (16,)))
 
-    x = jax.random.normal(next(keys), (batch, n, irreps.dim)) * 5 + 10.0
+    x = e3nn.normal(irreps, next(keys), (batch, n)) * 5
     x, state = b.apply(params, state, x)
 
     a = x.list[0]  # [batch, space, mul, 1]