+++ tests/exp_transformer_test.py
@@ -1,8 +1,7 @@
+import e3nn_jax as e3nn
+import haiku as hk
 import jax
 import jax.numpy as jnp
-import haiku as hk
-
-from e3nn_jax import IrrepsArray, radius_graph, spherical_harmonics, sus, soft_one_hot_linspace
 from e3nn_jax.experimental.transformer import Transformer
 
 
@@ -10,10 +9,10 @@ def test_transformer(keys):
     @hk.without_apply_rng
     @hk.transform
     def model(pos, src, dst, node_feat):
-        edge_attr = spherical_harmonics("0e + 1e + 2e", pos[dst] - pos[src], True)
+        edge_attr = e3nn.spherical_harmonics("0e + 1e + 2e", pos[dst] - pos[src], True)
         edge_distance = jnp.linalg.norm(pos[dst] - pos[src], axis=-1)
-        edge_weight_cutoff = sus(3.0 * (2.0 - edge_distance))
-        edge_scalar_attr = soft_one_hot_linspace(
+        edge_weight_cutoff = e3nn.sus(3.0 * (2.0 - edge_distance))
+        edge_scalar_attr = e3nn.soft_one_hot_linspace(
             edge_distance, start=0.0, end=2.0, number=5, basis="smooth_finite", cutoff=True
         )
 
@@ -32,8 +31,8 @@ def test_transformer(keys):
             [1.0, 0.0, 0.0],
         ]
     )
-    src, dst = radius_graph(pos, 2.0)
-    node_feat = IrrepsArray.randn("2x0e + 2x1e + 2x2e", next(keys), (pos.shape[0],))
+    src, dst = e3nn.radius_graph(pos, 2.0)
+    node_feat = e3nn.normal("2x0e + 2x1e + 2x2e", next(keys), (pos.shape[0],))
 
     w = model.init(next(keys), pos, src, dst, node_feat)
     apply(w, pos, src, dst, node_feat)